backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-openai
langchain-community
langchain-core
langchain-aws

pydantic
pydantic-settings
redis
matplotlib
pandas
yfinance
numpy
requests
python-multipart
aiohttp
GoogleNews
boto3
ta
youtube-transcript-api
google-api-python-client
google-auth-oauthlib 
google-auth-httplib2
aiohttp
asyncio
matplotlib-inline
nest-asyncio
pymongo
requests
beautifulsoup4
fake-useragent
tqdm
motor
